+++ packages/platform-sdk-btc-indexer/bin/meow.js
@@ -1,0 +1,36 @@
+#!/usr/bin/env node
+
+const meow = require("meow");
+
+module.exports = meow(
+	`
+	Usage
+	  $ pbtc <input>
+`,
+	{
+		flags: {
+			coin: {
+				type: "string",
+				alias: "c",
+				default: "btc",
+			},
+			network: {
+				type: "string",
+				alias: "c",
+				default: "mainnet",
+			},
+			host: {
+				type: "string",
+			},
+			username: {
+				type: "string",
+			},
+			password: {
+				type: "string",
+			},
+			database: {
+				type: "string",
+			},
+		},
+	},
+);